requirements.txt
Django==1.4.2
South
django-endless-pagination
sorl-thumbnail
django_compressor